Add Response model file for Character
