+++ rick-and-morty-app/rick-and-morty-app/API.swift
@@ -7,44 +7,6 @@
 
 import Foundation
 
-// DATA MODEL
-struct ResponseInfo: Codable {
-    let info: Info
-    let results: [Character]
-}
-
-struct Info: Codable {
-    let count: Int
-    let pages: Int
-    let next: String?
-    let prev: String?
-}
-
-struct Character: Codable {
-    let id: Int
-    let name: String
-    let status: String
-    let species: String
-    let type: String
-    let gender: String
-    let origin: Origin
-    let location: Location
-    let image: String
-    let episode: [String]
-    let url: String
-    let created: String
-}
-
-struct Origin: Codable {
-    let name: String
-    let url: String
-}
-
-struct Location: Codable {
-    let name: String
-    let url: String
-}
-
 let url = URL(string: "https://rickandmortyapi.com/api/character")
 
 /*
